EF: compile errors

+++ iModelBridgeCore/WSClient/DgnDbServerClient/DgnDbClient.cpp
@@ -266,7 +266,7 @@ DgnDbServerRepositoriesTaskPtr DgnDbClient::GetRepositories(ICancellationTokenPt
             }
 
         double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-        DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "Success.");
+        DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "Success.");
         return DgnDbServerRepositoriesResult::Success(repositories);
         });
     }
@@ -391,7 +391,7 @@ DgnDbPtr CleanDb(DgnDbCR db)
     tempdb->SaveChanges();
     //NEEDSWORK: end of file cleanup
     double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-    DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "");
+    DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "");
     return tempdb;
     }
 
@@ -480,7 +480,7 @@ DgnDbServerRepositoryTaskPtr DgnDbClient::CreateNewRepository(Dgn::DgnDbCR db, U
         })->Then<DgnDbServerRepositoryResult>([=]
             {
             double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-            DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "");
+            DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "");
             return *finalResult;
             });
     }
@@ -549,7 +549,7 @@ DgnDbServerBriefcaseTaskPtr DgnDbClient::OpenBriefcase(Dgn::DgnDbPtr db, bool do
                     if (result.IsSuccess())
                         {
                         double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-                        DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "");
+                        DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "");
                         finalResult->SetSuccess(briefcase);
                         }
                     else
@@ -562,7 +562,7 @@ DgnDbServerBriefcaseTaskPtr DgnDbClient::OpenBriefcase(Dgn::DgnDbPtr db, bool do
             else
                 {
                     double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-                    DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "");
+                    DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "");
                     finalResult->SetSuccess(briefcase);
                 }
             }
@@ -725,7 +725,7 @@ DgnDbServerBriefcaseInfoTaskPtr DgnDbClient::AcquireBriefcaseToDir(RepositoryInf
                     {
                     finalResult->SetSuccess(DgnDbServerBriefcaseInfo(filePath, isReadOnly));
                     double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-                    DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "Download successful.");
+                    DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "Download successful.");
                     }
                 else
                     {
@@ -787,7 +787,7 @@ DgnDbServerStatusTaskPtr DgnDbClient::DeleteRepository(RepositoryInfoCR reposito
         else
             {
             double end = BeTimeUtilities::GetCurrentTimeAsUnixMillisDouble();
-            DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, end - start, "Success.");
+            DgnDbServerLogHelper::Log(SEVERITY::LOG_INFO, methodName, (float)(end - start), "Success.");
             return DgnDbServerStatusResult::Success();
             }
         });
